requirements.txt
qdrant-client
sentence-transformers 
neo4j 
markdown
python-dotenv